requirements.txt
discord.py==2.1.1
Pillow==9.4.0
opencv-python==4.7.0.72
git+https://github.com/tdh8316/triangler/